OSSM-2376: Move kube controller to the federation package
